Health-web/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.1
mysql-connector-python==8.4.0
